Add positive specs for Socks5Handshake.parseRequest()

+++ src/main/com/github/c0va23/asyncSocksProxy/Exceptions.kt
@@ -7,7 +7,7 @@ class UnknownVersion(version: Byte) :
 class UnknownCommand(code: Byte) :
         SocksException("Unknown command code $code")
 
-class UnknownMethod(code: Short) :
+class UnknownMethod(code: Byte) :
         SocksException("Unknown method $code")
 
 class UnimplementedCommand(command: Command) :
